ml-service/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
python-multipart==0.0.20
spacy==3.8.4
scikit-learn==1.6.1
openai==1.59.9
python-docx==1.1.2
PyPDF2==3.0.1
httpx==0.28.1
structlog==24.4.0
